Add middleware which appends nosniff header
