+++ executor/api/rest/server.go
@@ -8,7 +8,6 @@ import (
 	"net/url"
 
 	"github.com/go-logr/logr"
-	guuid "github.com/google/uuid"
 	"github.com/gorilla/mux"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
@@ -21,15 +20,6 @@ import (
 	"github.com/seldonio/seldon-core/executor/predictor"
 	v1 "github.com/seldonio/seldon-core/operator/apis/machinelearning.seldon.io/v1"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
-)
-
-const (
-	CLOUDEVENTS_HEADER_ID_NAME             = "Ce-Id"
-	CLOUDEVENTS_HEADER_SPECVERSION_NAME    = "Ce-Specversion"
-	CLOUDEVENTS_HEADER_SOURCE_NAME         = "Ce-Source"
-	CLOUDEVENTS_HEADER_TYPE_NAME           = "Ce-Type"
-	CLOUDEVENTS_HEADER_PATH_NAME           = "Ce-Path"
-	CLOUDEVENTS_HEADER_SPECVERSION_DEFAULT = "0.3"
 )
 
 type SeldonRestApi struct {
@@ -159,42 +149,6 @@ func (r *SeldonRestApi) Initialise() {
 			r.Router.NewRoute().Path("/v1/models/{" + ModelHttpPathVariable + "}/metadata").Methods("GET").HandlerFunc(r.wrapMetrics(metric.MetadataHttpServiceName, r.metadata))
 		}
 	}
-}
-
-type CloudeventHeaderMiddleware struct {
-	deploymentName string
-	namespace      string
-}
-
-func (h *CloudeventHeaderMiddleware) Middleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Checking if request is cloudevent based on specname being present
-		if _, ok := r.Header[CLOUDEVENTS_HEADER_SPECVERSION_NAME]; ok {
-			puid := r.Header.Get(payload.SeldonPUIDHeader)
-			w.Header().Set(CLOUDEVENTS_HEADER_ID_NAME, puid)
-			w.Header().Set(CLOUDEVENTS_HEADER_SPECVERSION_NAME, CLOUDEVENTS_HEADER_SPECVERSION_DEFAULT)
-			w.Header().Set(CLOUDEVENTS_HEADER_PATH_NAME, r.URL.Path)
-			w.Header().Set(CLOUDEVENTS_HEADER_TYPE_NAME, "seldon."+h.deploymentName+"."+h.namespace+".response")
-			w.Header().Set(CLOUDEVENTS_HEADER_SOURCE_NAME, "seldon."+h.deploymentName)
-		}
-
-		next.ServeHTTP(w, r)
-	})
-}
-
-func puidHeader(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		puid := r.Header.Get(payload.SeldonPUIDHeader)
-		if len(puid) == 0 {
-			puid = guuid.New().String()
-			r.Header.Set(payload.SeldonPUIDHeader, puid)
-		}
-		if res_puid := w.Header().Get(payload.SeldonPUIDHeader); len(res_puid) == 0 {
-			w.Header().Set(payload.SeldonPUIDHeader, puid)
-		}
-
-		next.ServeHTTP(w, r)
-	})
 }
 
 func (r *SeldonRestApi) checkReady(w http.ResponseWriter, req *http.Request) {
